+++ test/integration/GCClassic/integrationTestCreate.sh
@@ -26,8 +26,11 @@
 thisDir=$(pwd -P)
 cd "${thisDir}"
 
+# Path to the test/shared folder in source code
+sharedDir=$(realpath "${thisDir}/../../shared")
+
 # Source the script containing utility functions and variables
-commonFuncs="${thisDir}/../../shared/commonFunctionsForTests.sh"
+commonFuncs="${sharedDir}/commonFunctionsForTests.sh"
 . "${commonFuncs}"
 
 #=============================================================================
@@ -103,6 +106,7 @@ execDir="${itRoot}/${EXEC_DIR}"
 logsDir="${itRoot}/${LOGS_DIR}"
 scriptsDir="${itRoot}/${SCRIPTS_DIR}"
 rundirsDir="${itRoot}/${RUNDIRS_DIR}"
+utilsDir="${itRoot}/${UTILS_DIR}"
 
 # Get absolute path of the environment file
 envFile=$(absolute_path "${envFile}")
@@ -149,17 +153,32 @@ ln -s "${superProjectDir}" ${itRoot}/CodeDir
 # Copy files to the proper folders
 #=============================================================================
 
-printf "\nCopying run scripts to: ${itRoot}/${SCRIPTS_DIR}\n"
+printf "\nCopying run scripts to:      ${scriptsDir}\n"
 cp -f ${thisDir}/integration*.sh     ${scriptsDir}
 cp -f ${commonFuncs}                 ${scriptsDir}
 cp -f ${thisDir}/README.md           ${scriptsDir}
 cp -f ${thisDir}/README.testroot.md  ${itRoot}/README.md
 
-# Only copy the environment file for Harvard Cannon
-[[ "X${site}" == "XCANNON" ]] && cp -f ${envFile} ${envDir}/gcclassic.env
+if [[ "X${site}" == "XCANNON" ]]; then
 
-# This is necessary on Compute1 to make all scripts executable
-chmod 755 -R ${scriptsDir}
+    # Copy Cannon environment file
+    cp -f  ${envFile} ${envDir}/gcclassic.env
+
+    # Copy Cannon utility scripts
+    printf "Copying utility scripts to   ${utilsDir}\n"
+    cp -fR ${sharedDir}/utils/cannon/integrationTest  ${utilsDir}
+
+elif [[ "X${site}" == "XCOMPUTE1" ]]; then
+
+    # Copy Compute1 utility scripts
+    printf "Copying utility scripts to   ${utilsDir}\n"
+    cp -fR ${sharedDir}/utils/compute1/integrationTest  ${utilsDir}
+
+    # Force scripts to be executable (Compute1 resets permissions)
+    chmod 755 -R ${scriptsDir}
+    chmod 755 -R ${utilsDir}
+
+fi
 
 # Log file with echoback from rundir creation
 log="${logsDir}/createIntegrationTests.log"
@@ -293,7 +312,9 @@ unset logsDir
 unset rundirsDir
 unset superProjectDir
 unset scriptsDir
+unset sharedDir
 unset thisDir
+unset utilsDir
 
 # Free imported variables
 unset FILL