+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - GCC/GCHP integration tests will exit immediately if `scheduler` is omitted.
 - Don't create run directories for integration/parallel tests if invoked with `-t compile`
 - Refactor integration and parallel test scripts to reduce the number of input arguments
+- Copy utility scripts that allow you to resubmit failed to integration and parallel test root directories
 
 ### Fixed
 - Fixed unit conversions in GEOS-only code